+++ go/libraries/doltcore/sqle/sqldelete_test.go
@@ -16,6 +16,7 @@ package sqle
 
 import (
 	"context"
+	"github.com/dolthub/dolt/go/libraries/doltcore/sqle/docsTable"
 	"testing"
 
 	sql "github.com/dolthub/go-mysql-server/sql"
@@ -23,7 +24,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/dolthub/dolt/go/libraries/doltcore/doltdb"
-	"github.com/dolthub/dolt/go/libraries/doltcore/doltdocs"
 	"github.com/dolthub/dolt/go/libraries/doltcore/dtestutils"
 	"github.com/dolthub/dolt/go/libraries/doltcore/schema"
 	. "github.com/dolthub/dolt/go/libraries/doltcore/sql/sqltestutil"
@@ -194,7 +194,7 @@ var systemTableDeleteTests = []DeleteTest{
 	{
 		Name: "delete dolt_docs",
 		AdditionalSetup: CreateTableFn("dolt_docs",
-			doltdocs.DoltDocsSchema,
+			docsTable.DoltDocsSchema,
 			NewRow(types.String("LICENSE.md"), types.String("A license"))),
 		DeleteQuery: "delete from dolt_docs",
 		ExpectedErr: "cannot delete from table",
